+++ hrs-portlets-api/src/main/java/edu/wisc/hr/dao/url/HrsUrlDao.java
@@ -46,6 +46,8 @@ public interface HrsUrlDao {
      */
     public static String APPROVE_PAYABLE_TIME_KEY = "Approve Payable time";
 
+    public static String SELF_SERVICE_DIRECT_DEPOSIT_KEY = "Direct Deposit";
+
     /**
      * The name of the key the value of which is the deep link to HRS for an
      * employee to view the garnishments on their own earnings.